+++ fabric/src/main/java/com/chefmooon/colourfulclocks/data/fabric/RecipeGenerator.java
@@ -31,11 +31,23 @@ public class RecipeGenerator extends FabricRecipeProvider {
         buildPocketWatchRecipe(ColourfulClocksItemsImpl.DIAMOND_POCKET_WATCH.get(), PocketWatchTypes.DIAMOND, recipeOutput);
         buildPocketWatchRecipe(ColourfulClocksItemsImpl.NETHERITE_POCKET_WATCH.get(), PocketWatchTypes.NETHERITE, recipeOutput);
 
+        buildPocketWatchRecipe(ColourfulClocksItemsImpl.QUARTZ_POCKET_WATCH.get(), PocketWatchTypes.QUARTZ, recipeOutput);
+        buildPocketWatchRecipe(ColourfulClocksItemsImpl.AMETHYST_POCKET_WATCH.get(), PocketWatchTypes.AMETHYST, recipeOutput);
+        buildPocketWatchRecipe(ColourfulClocksItemsImpl.LAPIS_LAZULI_POCKET_WATCH.get(), PocketWatchTypes.LAPIS_LAZULI, recipeOutput);
+        buildPocketWatchRecipe(ColourfulClocksItemsImpl.REDSTONE_POCKET_WATCH.get(), PocketWatchTypes.REDSTONE, recipeOutput);
+        buildPocketWatchRecipe(ColourfulClocksItemsImpl.EMERALD_POCKET_WATCH.get(), PocketWatchTypes.EMERALD, recipeOutput);
+
         buildPendulumRecipe(ColourfulClocksItemsImpl.IRON_PENDULUM.get(), PendulumTypes.IRON, recipeOutput);
         buildPendulumRecipe(ColourfulClocksItemsImpl.COPPER_PENDULUM.get(), PendulumTypes.COPPER, recipeOutput);
         buildPendulumRecipe(ColourfulClocksItemsImpl.GOLD_PENDULUM.get(), PendulumTypes.GOLD, recipeOutput);
         buildPendulumRecipe(ColourfulClocksItemsImpl.DIAMOND_PENDULUM.get(), PendulumTypes.DIAMOND, recipeOutput);
         buildPendulumRecipe(ColourfulClocksItemsImpl.NETHERITE_PENDULUM.get(), PendulumTypes.NETHERITE, recipeOutput);
+
+        buildPendulumRecipe(ColourfulClocksItemsImpl.QUARTZ_PENDULUM.get(), PendulumTypes.QUARTZ, recipeOutput);
+        buildPendulumRecipe(ColourfulClocksItemsImpl.AMETHYST_PENDULUM.get(), PendulumTypes.AMETHYST, recipeOutput);
+        buildPendulumRecipe(ColourfulClocksItemsImpl.LAPIS_LAZULI_PENDULUM.get(), PendulumTypes.LAPIS_LAZULI, recipeOutput);
+        buildPendulumRecipe(ColourfulClocksItemsImpl.REDSTONE_PENDULUM.get(), PendulumTypes.REDSTONE, recipeOutput);
+        buildPendulumRecipe(ColourfulClocksItemsImpl.EMERALD_PENDULUM.get(), PendulumTypes.EMERALD, recipeOutput);
 
         buildBornholmRecipes(recipeOutput);
     }
@@ -46,11 +58,11 @@ public class RecipeGenerator extends FabricRecipeProvider {
                 .pattern(" C ")
                 .pattern("ABA")
                 .pattern(" A ")
-                .define('A', pocketWatchTypes.getItem())
+                .define('A', pocketWatchTypes.getCraftingIngredient())
                 .define('B', Items.REDSTONE)
                 .define('C', Items.QUARTZ)
                 .unlockedBy("has_any_ingredient", RecipeProvider.inventoryTrigger(ItemPredicate.Builder.item().of(
-                        pocketWatchTypes.getItem(),
+                        pocketWatchTypes.getCraftingIngredient(),
                         Items.REDSTONE,
                         Items.QUARTZ)))
                 .save(recipeOutput, RecipeProvider.getSimpleRecipeName(item));
@@ -58,9 +70,9 @@ public class RecipeGenerator extends FabricRecipeProvider {
 
     private static void buildPendulumRecipe(ItemLike item, PendulumTypes pendulumTypes, RecipeOutput recipeOutput) {
         ShapedRecipeBuilder.shaped(RecipeCategory.MISC, item)
-                .pattern("B")
-                .pattern("A")
-                .pattern("A")
+                .pattern("  B")
+                .pattern("AA ")
+                .pattern("AA ")
                 .define('A', pendulumTypes.getCraftingIngredient())
                 .define('B', Items.CHAIN)
                 .unlockedBy("has_any_ingredient", RecipeProvider.inventoryTrigger(ItemPredicate.Builder.item().of(