+++ changelog.md
@@ -1,5 +1,7 @@
 ### Added
 - Now available for Fabric!
+- New Pocket Watch and Pendulum Types
+  - Emerald, Amethyst, Quartz, Lapis Lazuli, Redstone
 
 ### Updated
 - Pocket Watches can now be closed and opened
@@ -8,5 +10,10 @@
   - Hand now moves properly when interacting with the door
   - Door will now open/close when struck by a wind charge
 - Pendulum Updates
+  - Recipe Changed: now uses 1x Iron Chain, 4x Item
   - Each pendulum sound is now separate and can be set with a datapack
   - New sounds will be added soon
+- Textures Updated
+  - Iron Pocket Watch
+  - Clock Hands -> Copper, Gold, Diamond, Netherite
+- Creative Tab Icon is now the Netherite Pocket Watch